+++ test_HW3/(переименовать в случае необходимости)conftest.py
